chore: moves code to separate code fiel instead of routes file
